# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TrustLoopGuard API
  description: Real-time guardrail runtime for AI agents.
  license:
    name: Apache-2.0
  version: 0.0.1
paths:
...
  /v1/label-policies:
    get:
      tags:
      - label-policies
      summary: |-
        `GET /v1/label-policies`. Returns all active (non-deleted) policies
        for the workspace, including disabled ones.
      operationId: list_label_policies
      responses:
        '200':
          description: All label policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceLabelPolicyListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - label-policies
      summary: '`POST /v1/label-policies` — upsert a per-origin override.'
      operationId: upsert_label_policy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertSourceLabelPolicyRequest'
        required: true
      responses:
        '201':
          description: Label policy created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceLabelPolicyEntry'
        '400':
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Label policy failed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/label-policies/{origin}:
    get:
      tags:
      - label-policies
      summary: '`GET /v1/label-policies/:origin`.'
      operationId: get_label_policy
      parameters:
      - name: origin
        in: path
        description: Source origin (snake_case)
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Label policy found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceLabelPolicyEntry'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: No policy for this origin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Unknown origin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - label-policies
      summary: '`DELETE /v1/label-policies/:origin`. Soft-delete via the store.'
      operationId: delete_label_policy
      parameters:
      - name: origin
        in: path
        description: Source origin (snake_case)
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Label policy deleted
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: No policy for this origin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Unknown origin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1/policies/validate:
    post:
      tags:
      - policies
      summary: '`POST /v1/policies/validate` — validate policy YAML or JSON without saving it.'
      operationId: validate_policy
      requestBody:
        description: Policy document, YAML or JSON
        content:
          application/yaml:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyValidateResponse'
        '400':
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
...
    ApiError:
      type: object
      description: |-
        Canonical error envelope returned on non-2xx responses from every
        TrustLoopGuard endpoint. SDKs deserialize this body to produce typed
        errors; integrators don't have to inspect status codes by hand.

        The shape is intentionally minimal — `code` drives SDK-side fan-out,
        `message` is for logs, `retriable` tells callers whether the same
        request may be retried, and `details` is opaque so the server can
        add validation field paths without breaking SDKs.
      required:
      - code
      - message
      - retriable
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        details:
          description: |-
            Opaque structured details (e.g. validation field path).
            Defaults to `null`; servers may add fields without breaking SDKs.
        message:
          type: string
        retriable:
          type: boolean
          description: |-
            Whether the caller may retry the same request without modification.
            SDKs honor `Retry-After` when present in addition to this flag.
    ApiErrorCode:
      type: string
      description: |-
        Stable error code dictionary. Add variants here when introducing new
        failure modes; never repurpose an existing variant — SDK callers may
        be branching on it.
      enum:
      - invalid
      - unauthorized
      - forbidden
      - not_found
      - gone
      - unprocessable
      - rate_limited
      - internal
      - unavailable
...
    Confidentiality:
      type: string
      enum:
      - public
      - private
      - secret
      - identity
      - unknown
...
    Integrity:
      type: string
      enum:
      - low
      - medium
      - high
      - unknown
...
    Origin:
      type: string
      enum:
      - user
      - system
      - tool
      - memory
      - file
      - web
      - email
      - api
      - unknown
...
    PolicyValidateResponse:
      type: object
      required:
      - valid
      - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/PolicyValidationIssue'
        policy_id:
          type:
          - string
          - 'null'
        valid:
          type: boolean
    PolicyValidationIssue:
      type: object
      required:
      - path
      - message
      properties:
        message:
          type: string
        path:
          type: string
...
    SourceLabelPolicy:
      type: object
      description: |-
        One workspace override row: per-origin label defaults. A family left
        `None` inherits the built-in origin default.
      required:
      - origin
      properties:
        confidentiality:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Confidentiality'
        integrity:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Integrity'
        origin:
          $ref: '#/components/schemas/Origin'
        trust:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Trust'
    SourceLabelPolicyEntry:
      type: object
      description: |-
        A registry row as seen by the control plane: the wire policy plus
        its `enabled` flag. Disabled policies stay manageable but are
        skipped during runtime label resolution.
      required:
      - policy
      - enabled
      properties:
        enabled:
          type: boolean
        policy:
          $ref: '#/components/schemas/SourceLabelPolicy'
    SourceLabelPolicyListResponse:
      type: object
      required:
      - policies
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/SourceLabelPolicyEntry'
...
    Trust:
      type: string
      enum:
      - trusted
      - untrusted
      - unknown
...
    UpsertSourceLabelPolicyRequest:
      allOf:
      - $ref: '#/components/schemas/SourceLabelPolicy'
      - type: object
        properties:
          enabled:
            type: boolean
      description: |-
        Upsert body for `POST /v1/label-policies`: the wire policy plus an
        optional `enabled` flag (defaults to `true`).
      anyOf:
      - required:
        - trust
      - required:
        - confidentiality
      - required:
        - integrity
...